Change the dnsmasq DNS domain to test-cluster.redhat.com
We want hosts in this network to be on the test-cluster.redhat.com
domain, so it makes sense to do this. However, it also means that
the 'localOnly' configuration only applies to names under this domain
and queries for e.g. registry.connect.redhat.com will be forwarded
to upstream DNS servers.

i.e. before:

 $> host registry.connect.redhat.com 192.168.126.1
 Using domain server:
 Name: 192.168.126.1
 Address: 192.168.126.1#53
 Aliases:

 Host registry.connect.redhat.com not found: 3(NXDOMAIN)

and after:

 $> host registry.connect.redhat.com 192.168.126.1
 Using domain server:
 Name: 192.168.126.1
 Address: 192.168.126.1#53
 Aliases:

 registry.connect.redhat.com is an alias for registry.connect.redhat.com2.edgekey.net.
 registry.connect.redhat.com2.edgekey.net is an alias for registry.connect.redhat.com2.edgekey.net.globalredir.akadns.net.
 registry.connect.redhat.com2.edgekey.net.globalredir.akadns.net is an alias for e40408.d.akamaiedge.net.
 e40408.d.akamaiedge.net has address 23.75.23.179
 e40408.d.akamaiedge.net has address 23.75.23.95