Add option to add custom manifests
Also added makefile target to dump ignition into a directory so it can
be inspected.

Also made registry-config.json .PHONY because it was really annoying to
change the pull secret and have those changes ignored. This is at the
cost of waiting 5 seconds for the installer to get pulled out of the
image each time, I think it's worth it though.